+++ ck.json
@@ -17,7 +17,6 @@
             "restaruranstImg": "https://blog.kakaocdn.net/dn/exLM42/btqwP7n6Me6/m2gfQtrk82t7iFVZkMLLH0/img.jpg"
         },
         {
-            "restaruranstTitle": "60계",
             "restaruranstImg": "https://blog.kakaocdn.net/dn/bYxk8F/btqzbVdZXnU/eDGWNmlponKBAabI83n5K0/img.jpg"
         },
         {
