+++ package.json
@@ -10,7 +10,6 @@
     "immer": "^9.0.12",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-
     "react-redux": "^7.2.8",
     "react-router-dom": "5.2.1",
     "react-scripts": "5.0.0",
@@ -19,7 +18,6 @@
     "redux-actions": "^2.6.5",
     "redux-thunk": "^2.4.1",
     "slick-carousel": "^1.8.1",
-
     "styled-components": "^5.3.5",
     "web-vitals": "^2.1.4"
   },
